==> 9_Information_search/1. Collect data.ipynb <==
Создайте папку data, а в ней папки page_texts и raw_pages

!pip install beautifulsoup4

import requests
import re
from bs4 import BeautifulSoup

USER_AGENT = 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_9_3) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/35.0.1916.47 Safari/537.36'

def get_page_text(url, user_agent, error_log=True):
    headers = {'User-Agent': user_agent}
    response = requests.get(url, headers=headers, allow_redirects=True)

    if response.status_code == 200:
        page_text = response.text
        return page_text
    elif error_log:
        print(f'Ошибка! Response code: {response.status_code}')
        print(f'URL: {url}')
        print('--------------')
    
    return None

index = 1
url = f'https://www.kinopoisk.ru/media/?page={index}'

page_text = get_page_text(url, USER_AGENT)

url

print(page_text[:1000])

from os import makedirs

soup = BeautifulSoup(page_text)
posts_list = soup.find_all('div', {'class': 'posts-grid__main-section-column'})
for p in posts_list:
    lnk = p.find('a').attrs['href']
    title = re.sub('[^А-ЯЁа-яё0-9\s]', ' ', p.text)
    title = re.sub('\s\s+', ' ', title)
    
    page_url = 'https://www.kinopoisk.ru' + lnk
    clean_path = '/'.join([d for d in page_url.split('/')[2:] if len(d) > 0])
    
    page_text = get_page_text(page_url, USER_AGENT)
    
    if page_text is None:
        continue
        
    dir_path = 'data/raw_pages/' + '/'.join(clean_path.split('/')[:-1])
    makedirs(dir_path, exist_ok=True)
    with open(dir_path + '/' + clean_path.split('/')[-1] + '.html', 'w', encoding='utf-8') as f:
        f.write(page_text)
            
#     print(page_url, title)

main_url = f'https://www.kinopoisk.ru/'

page_text = get_page_text(main_url, USER_AGENT)
soup = BeautifulSoup(page_text)

def clear_link(page_link):
    page_link = re.sub('^\/', '', page_link)
    page_link = re.sub('[\#\?].+', '', page_link)
    
    return page_link

links = set()
for a in soup.find_all('a'):
    if not a.has_attr('href'):
        continue
    
    page_link = a.attrs['href']
    
    if re.search('^(http)', page_link):
        if main_url not in page_link or main_url == page_link:
            continue
    
    page_link = clear_link(page_link)
    
    if len(page_link) == 0:
        continue 
    
    page_link = main_url + page_link
    
    links.add(page_link)
#     print(page_link)

links

def get_page_links(soup, main_url):
    links = set()
    for a in soup.find_all('a'):
        if not a.has_attr('href'):
            continue

        page_link = a.attrs['href']

        if re.search('^(http)', page_link):
            if main_url not in page_link or main_url == page_link:
                continue

        page_link = clear_link(page_link)

        if len(page_link) == 0:
            continue 

        page_link = main_url + page_link

        links.add(page_link)
    
    return links

get_page_links(soup, main_url)

# Сбор всех ссылок с ресурса
main_url = f'https://www.kinopoisk.ru/'

page_text = get_page_text(main_url, USER_AGENT)
soup = BeautifulSoup(page_text)

all_links = get_page_links(soup, main_url)
process_links = all_links

i = 0 # Для мастер-класса
while True:
    new_links = set()

    for lnk in process_links:
        page_text = get_page_text(lnk, USER_AGENT)
        
        if 
        soup = BeautifulSoup(page_text)
        links = get_page_links(soup, main_url)
        links = links.difference(all_links, new_links)

        new_links = new_links.union(links)

    if len(new_links) == 0:
        break

    all_links = all_links.union(new_links)
    process_links = new_links
    
    # Для мастер-класса
    i += 1
    if i == 10:
        break

len(all_links)
