==> 9_Information_search/2. Prepare data.ipynb <==
from bs4 import BeautifulSoup
from os import walk

import difflib
import re

pages_list = []

for dirpath, dirnames, filenames in walk('data/raw_pages'):
    if '.ipynb_checkpoints' in dirpath:
        continue
    
    dirpath = dirpath.replace('\\', '/') # для Windows
    for fn in filenames:
        if '.DS_Store' in fn:
            continue 
            
        fp = f'{dirpath}/{fn}'
        pages_list.append(fp)

pages_list[:10]

def remove_script(file):
    soup = BeautifulSoup(''.join(file), 'html.parser')
    for s in soup.select('script'):
        s.extract()
        
    for f in soup.select('footer'):
        f.extract()
    
    for f in soup.select('header'):
        f.extract()
        
    return str(soup).split('\n')

fp_1 = 'data/raw_pages/www.kinopoisk.ru/media/video/4005161.html'
fp_2 = 'data/raw_pages/www.kinopoisk.ru/media/video/4005135.html'

with open(fp_1, 'r') as f:
    file1 = remove_script(f.readlines())
with open(fp_2, 'r') as f:
    file2 = remove_script(f.readlines())

def clean_diff(diff):
    diff = re.sub('\<[^\<\>]+\>', ' ', diff)
    diff = re.sub('&nbsp;', ' ', diff)
    diff = re.sub('\xa0', ' ', diff)
    diff = re.sub('\s\s+', ' ', diff)
    diff = re.sub('^[\+\-] ', '', diff)
    
    return diff

page_lines = []
for diff in difflib.ndiff(file1, file2):
    if re.search('^\+ ', diff) is None:
        continue
        
    diff = clean_diff(diff)
    
    if len(diff) == 0:
        continue
    
    page_lines.append(diff)
#     print(diff)

page_text = ' '.join(page_lines)

print(page_text)

def extract_page_text(file, origin_file):
    
    page_lines = []
    for diff in difflib.ndiff(origin_file, file):
        if re.search('^\+ ', diff) is None:
            continue

        diff = clean_diff(diff)

        if len(diff) == 0:
            continue

        page_lines.append(diff)

    page_text = '. '.join(page_lines)
    
    return page_text

# Извлекаем тексты, и сохраняем в файлы

page_texts_path = 'data/page_texts/'
create_fp = lambda fp: fp.replace('/', '+').replace('\\', '+') + '.txt'


with open(pages_list[0], 'r', encoding="utf-8") as f:
    origin_file = remove_script(f.readlines())

with open(pages_list[1], 'r', encoding="utf-8") as f:
    file = remove_script(f.readlines())

page_text = extract_page_text(origin_file, file)

with open(page_texts_path + create_fp(pages_list[0]), 'w', encoding="utf-8") as f:
    f.write(page_text)

for fp in pages_list[1:]:
    with open(fp, 'r', encoding="utf-8") as f:
        file = remove_script(f.readlines())
    
    page_text = extract_page_text(file, origin_file)
    with open(page_texts_path + create_fp(fp), 'w', encoding="utf-8") as f:
        f.write(page_text)

pages_texts = {}

for dirpath, dirnames, filenames in walk(page_texts_path):
    if '.ipynb_checkpoints' in dirpath:
        continue
        
    for fn in filenames:
        if '.DS_Store' in fn:
            continue
            
        fp = f'{dirpath}/{fn}'
        with open(fp, 'r', encoding="utf-8") as f:
            pt = ' '.join(f.readlines())
        
        original_link = fp.split('/')[-1].replace('|', '/').replace('.txt', '')
        pages_texts[original_link] = pt

!pip install pymorphy2
!pip install pymorphy2-dicts

# ставим либо DAWG-Python, либо DAWG (он быстрее)
!pip install DAWG-Python

# Если есть Cython и компилятор в системе
# !pip install DAWG

import nltk
import pymorphy2
from string import punctuation

print(punctuation)

MORPH = pymorphy2.MorphAnalyzer()

def preprocess_text(text):
    text = text.lower()
    text = re.sub('\s\s+', ' ', text)
    
    return text

def extract_sentences(text):
    text = preprocess_text(text)

    sentences = []
    for s in nltk.sent_tokenize(text):
        wrds = []
        for wrd in nltk.word_tokenize(s):
            if wrd in punctuation:
                continue
            
            wrd = MORPH.parse(wrd)[0].normal_form
            wrds.append(wrd)
            
        clear_sentence = ' '.join(wrds)
        
        sentences.append(clear_sentence)
    
    return sentences

MORPH.parse('мамой')[0]

nltk.sent_tokenize('я люблю тебя до слёз. будь счастлива!')

nltk.word_tokenize('мама, мыла раму папу-макаку за 170,00 рублей 170рублей')

import pandas as pd
import numpy as np

len(pages_texts)

nltk.download('punkt')

full_link_df = pd.DataFrame(columns=['link', 'sentence_order', 'sentence_text'])

for link, text in pages_texts.items():
    sentences = extract_sentences(text)
    sentence_count = len(sentences)
    
    link_df = pd.DataFrame({'link': np.repeat(link, sentence_count),
                            'sentence_order': np.arange(sentence_count),
                            'sentence_text': sentences})
    
    full_link_df = full_link_df.append(link_df)

full_link_df

full_link_df.to_csv('data/link_page_sentences.csv', index=False)

pd.read_csv('data/link_page_sentences.csv').head(5)

!pip install pyarrow

full_link_df.reset_index().to_feather('data/link_page_sentences.ftr')

pd.read_feather('data/link_page_sentences.ftr').head(5)

full_link_df.to_parquet('data/link_page_sentences.parquet')

pd.read_parquet('data/link_page_sentences.parquet').head(5)
