==> 9_Information_search/3. Search engine.ipynb <==
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.decomposition import PCA
from sklearn.neighbors import NearestNeighbors

import pandas as pd

import nltk
from nltk.corpus import stopwords

nltk.download('stopwords')

# pages_data = pd.read_feather('data/link_page_sentences.ftr')
pages_data = pd.read_csv('data/link_page_sentences.csv')

pages_data.dropna(inplace=True)
# pages_data.drop('index', axis=1, inplace=True)
pages_data.set_index(['link', 'sentence_order'], inplace=True)

pages_data.head(5)

stopwords.words('russian')[:10]

pages_data.shape

# vectorizer = TfidfVectorizer(stop_words=stopwords.words('russian'), min_df=0.01)
vectorizer = TfidfVectorizer(stop_words=stopwords.words('russian'))
page_indices_array = vectorizer.fit_transform(pages_data['sentence_text']).toarray()
columns = list(vectorizer.vocabulary_.keys())

page_indices = pd.DataFrame(page_indices_array, columns=columns)
page_indices.index = pages_data.index
page_indices.shape

page_indices.head(5)

ranker = NearestNeighbors(n_neighbors=100, metric='cosine', n_jobs=-1)
ranker.fit(page_indices)

pages_data.values[0][0]

import re
import pymorphy2

from string import punctuation

MORPH = pymorphy2.MorphAnalyzer()

def preprocess_query(query):
    query = query.lower()
    query = re.sub('\s\s+', ' ', query)
    
    wrds = []
    for wrd in nltk.word_tokenize(query):
        if wrd in punctuation:
            continue

        wrd = MORPH.parse(wrd)[0].normal_form
        wrds.append(wrd)

    clear_query = ' '.join(wrds)
    
    return clear_query

query = 'Портрет Незнакомца фильм 2021'
query

preprocess_query(query)

query = preprocess_query(query)
query_vect = vectorizer.transform([query]).toarray()

scores, indices = ranker.kneighbors(query_vect)
scores = scores[0]
indices = indices[0]

scores, indices

page_indices.iloc[indices]

candidate_links = page_indices.iloc[indices].reset_index()['link']
score_df = pd.DataFrame({'candidate_link': candidate_links, 'score': scores})

score_df.head(5)

# group_score_df.groupby('candidate_link').mean()['score']
# group_score_df = score_df.groupby('candidate_link').min()['score']
group_score_df = score_df.groupby('candidate_link').sum()['score']
group_score_df = group_score_df.sort_values()

best_link = group_score_df.index[0]
best_link

group_score_df[:10].index.tolist()

pca = PCA(n_components=60)
pca_indices_array = pca.fit_transform(page_indices_array)

pca_indices_array[0]

ranker_pca = NearestNeighbors(n_neighbors=100, metric='cosine', n_jobs=-1)
ranker_pca.fit(pca_indices_array)

pca.transform(vectorizer.transform([query]).toarray())

def search_page(query, vectorizer, pca, ranker, page_count=10, score_type='sum'):
    query = preprocess_query(query)
    query_vect = vectorizer.transform([query]).toarray()
    query_vect = pca.transform(query_vect)
    
    scores, indices = ranker.kneighbors(query_vect)
    scores = scores[0]
    indices = indices[0]
    
    candidate_links = page_indices.iloc[indices].reset_index()['link']
    score_df = pd.DataFrame({'candidate_link': candidate_links, 'score': scores})
    
    if score_type == 'mean':
        group_score_df = group_score_df.groupby('candidate_link').mean()['score']
        group_score_df = group_score_df.sort_values(ascending=True)
    elif score_type == 'min':
        group_score_df = score_df.groupby('candidate_link').min()['score']
        group_score_df = group_score_df.sort_values(ascending=True)
    else:
        score_df['score'] = 1 - score_df['score']
        group_score_df = score_df.groupby('candidate_link').sum()['score']
        group_score_df = group_score_df.sort_values(ascending=False)
    
    return group_score_df[:page_count].index.tolist()

search_page('кинотавр фотографии', vectorizer, pca, ranker_pca, page_count=15, score_type='sum')

import pickle

with open('search_model', 'wb') as f:
    pickle.dump({'vectorizer': vectorizer, 'pca': pca, 'ranker': ranker_pca}, f)

with open('search_model', 'rb') as f:
    search_model = pickle.load(f)
    
search_model

vectorizer = search_model['vectorizer']
pca = search_model['pca']
ranker_pca = search_model['ranker']

vectorizer, pca, ranker_pca

search_page('дюна 2021', vectorizer, pca, ranker_pca, page_count=15, score_type='sum')
